pyproject.toml
[project]
name = "airas-experiments"
version = "0.1.0"
description = "AI Research Agent System - DiVA-LtM prompt tuning experiments"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
    "accelerate>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"